fix: make itemCount 64bit

+++ pkg/analyze/dir_test.go
@@ -35,7 +35,7 @@ func TestAnalyzeDir(t *testing.T) {
 	// test dir info
 	assert.Equal(t, "test_dir", dir.Name)
 	assert.Equal(t, int64(7+4096*3), dir.Size)
-	assert.Equal(t, 5, dir.ItemCount)
+	assert.Equal(t, int64(5), dir.ItemCount)
 	assert.True(t, dir.IsDir())
 
 	// test dir tree
@@ -76,7 +76,7 @@ func TestIgnoreDir(t *testing.T) {
 	).(*Dir)
 
 	assert.Equal(t, "test_dir", dir.Name)
-	assert.Equal(t, 1, dir.ItemCount)
+	assert.Equal(t, int64(1), dir.ItemCount)
 }
 
 func TestFlags(t *testing.T) {
@@ -99,7 +99,7 @@ func TestFlags(t *testing.T) {
 	sort.Sort(sort.Reverse(dir.Files))
 
 	assert.Equal(t, int64(28+4096*4), dir.Size)
-	assert.Equal(t, 7, dir.ItemCount)
+	assert.Equal(t, int64(7), dir.ItemCount)
 
 	// test file3
 	assert.Equal(t, "nested", dir.Files[0].GetName())
@@ -125,7 +125,7 @@ func TestHardlink(t *testing.T) {
 	dir.UpdateStats(make(fs.HardLinkedItems))
 
 	assert.Equal(t, int64(7+4096*3), dir.Size) // file2 and file3 are counted just once for size
-	assert.Equal(t, 6, dir.ItemCount)          // but twice for item count
+	assert.Equal(t, int64(6), dir.ItemCount)          // but twice for item count
 
 	// test file3
 	assert.Equal(t, "file3", dir.Files[0].(*Dir).Files[1].GetName())
@@ -154,7 +154,7 @@ func TestFollowSymlink(t *testing.T) {
 	sort.Sort(sort.Reverse(dir.Files))
 
 	assert.Equal(t, int64(9+4096*4), dir.Size)
-	assert.Equal(t, 7, dir.ItemCount)
+	assert.Equal(t, int64(7), dir.ItemCount)
 
 	// test file3
 	assert.Equal(t, "nested", dir.Files[0].GetName())
@@ -191,7 +191,7 @@ func TestGitAnnexSymlink(t *testing.T) {
 	sort.Sort(sort.Reverse(dir.Files))
 
 	assert.Equal(t, int64(967858083+7+4096*4), dir.Size)
-	assert.Equal(t, 7, dir.ItemCount)
+	assert.Equal(t, int64(7), dir.ItemCount)
 
 	// test file3
 	assert.Equal(t, "nested", dir.Files[0].GetName())
@@ -223,7 +223,7 @@ func TestBrokenSymlinkSkipped(t *testing.T) {
 	sort.Sort(sort.Reverse(dir.Files))
 
 	assert.Equal(t, int64(7+4096*4), dir.Size)
-	assert.Equal(t, 6, dir.ItemCount)
+	assert.Equal(t, int64(6), dir.ItemCount)
 
 	assert.Equal(t, '!', dir.Files[0].GetFlag())
 }
