+++ pkg/analyze/parallel.go
@@ -247,7 +247,7 @@ func (a *ParallelAnalyzer) processDir(path string) *Dir {
 
 	a.progressChan <- common.CurrentProgress{
 		CurrentItemName: path,
-		ItemCount:       len(files),
+		ItemCount:       int64(len(files)),
 		TotalSize:       totalSize,
 	}
 	return dir
